requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
python-dateutil>=2.8.0
apify-client>=1.6.0
python-dotenv>=1.0.0